arm64: dts: ti: k3-am62-verdin-dahlia: support sleep-moci
Previously, we had the sleep-moci pin set to always on. However, the
Dahlia carrier board supports disabling the sleep-moci when the system
is suspended to power down peripherals that support it. This reduces
overall power consumption. This commit adds support for this feature by
disabling the reg_force_sleep_moci regulator and adding a new regulator
for the USB hub that can be turned off when the system is suspended.

Upstream-Status: Backport [3935fbc87ddebea5439f3ab6a78b1e83e976bf88]
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>