arm64: dts: ti: k3-am69-aquila: remove interrupt from rx8130 rtc
Remove interrupt from the rx8130 RTC as the driver incorrectly reports
that the update interrupt is working.

The update interrupt should send a request once per second. This is
achieved by setting an alarm triggering the interrupt on the next
second tick. However, the rx8130 RTC has a one-minute resolution alarm.

Upstream-Status: Pending
Signed-off-by: Emanuele Ghidoli <emanuele.ghidoli@toradex.com>