arm64: dts: ti: am62p-verdin: add eeprom aliases
Add eeprom aliases, eeprom0 is the on-som eeprom, eeprom1 is the
on-carrier eeprom. Those aliases are used in U-Boot and defining those
in the Linux DT allows to have one less difference between the twos with
no real drawback (a marginally bigger dtb in Linux makes no real
difference).

Upstream-Status: Pending
Signed-off-by: Francesco Dolcini <francesco.dolcini@toradex.com>