arm64: dts: ti: k3-am69-aquila-dev: add Microphone Jack to sound card
The simple-audio-card's microphone widget currently connects to the
headphone jack. Routing the microphone input to the microphone jack
allows for independent operation of the microphone and headphones.

This resolves the following boot-time kernel log message, which
indicated a conflict when the microphone and headphone functions were
not separated:
  debugfs: File 'Headphone Jack' in directory 'dapm' already present!

Upstream-Status: Pending
Fixes: 64921c121ca8 ("arm64: dts: ti: k3-am69-aquila-dev: Add audio codec")
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>