arm64: dts: ti: k3-am62p-verdin-dahlia: add Microphone Jack to sound card
The simple-audio-card's microphone widget currently connects to the
headphone jack. Routing the microphone input to the microphone jack
allows for independent operation of the microphone and headphones.

This resolves the following boot-time kernel log message, which
indicated a conflict when the microphone and headphone functions were
not separated:
  debugfs: File 'Headphone Jack' in directory 'dapm' already present!

Upstream-Status: Pending
Fixes: 0c1e7b62a8bf ("arm64: dts: ti: am62p-verdin: add dahlia")
Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>