FROMLIST: net: ti: icssg-prueth: Use page_pool API for RX buffer allocation
This is to prepare for native XDP support.

The page pool API is more faster in allocating pages than
__alloc_skb(). Drawback is that it works at PAGE_SIZE granularity
so we are not efficient in memory usage.
i.e. we are using PAGE_SIZE (4KB) memory for 1.5KB max packet size.

This is a cherry-pick of commit b95f4a8493ac ("FROMLIST: net: ti:
icssg-prueth: Use page_pool API for RX buffer allocation") from TI's
downstream ti-linux-v6.12 branch. It adds the necessary changes for
icssg_prueth_sr1 to fix a build error related to function
`prueth_dma_rx_push` that was changed with the bump to stable v6.6.84
and the rest of the page_pool API changes that were applied to
icssg_prueth_sr1. If we migrate to ti-linux-v6.12 in the future, this
cherry-pick won't be needed.

Upstream-Status: Inappropriate [other]
Fixes: 8b467fa3694a ("Merge tag 'v6.6.84'")
Link: https://lore.kernel.org/all/20250305101422.1908370-2-m-malladi@ti.com/
Signed-off-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: MD Danish Anwar <danishanwar@ti.com>
Signed-off-by: Meghana Malladi <m-malladi@ti.com>