i2c: omap: Increase timeout to 5 seconds
When adding the gmsl serializer, the I2C bus is timing out for other
devices on the bus. Increasing the timeout to 5 seconds solved the
poblem.

Don't know why it is happening, maybe the bus is being held for longer than 1
sec to communicate with the serializier over the gmsl link. Need to investigate
more. For now, I will go foward to make the whole system to work.

Upstream-Status: Pending
Signed-off-by: João Paulo Gonçalves <jpaulo.silvagoncalves@gmail.com>