media: cadence: csi2rx: Support reading link frequency from mbus_config
Newer drivers are using get_mbus_config to pass the link frequency and
not the control V4L2_CID_LINK_FREQ. Add the to the driver the
possibilitie to handle this newer drivers too for reading the link
frequency. For now, this seem an issue on the downstream driver as
upstream is fixed in e7bad98c205d17c745de9d83ebf73e53cbf99d48.

Upstream-Status: Innapropriate [other].
Signed-off-by: João Paulo Gonçalves <jpaulo.silvagoncalves@gmail.com>